Allow inheritance as microflow inputparameter
Currently it is not allowed to use a specialization of Queue.Job as input parameter in your microflow. This makes it harder if you want to link your job to input parameters. You are required to create an association from the job entity or have Job as your input parameter and then cast it to the specified job in your microflow, which feels foolish because i already know the input type of the specific microflow.