Allow inheritance as microflow inputparameter (#5)
* Allow inheritance as microflow inputparameter

Currently it is not allowed to use a specialization of Queue.Job as input parameter in your microflow. This makes it harder if you want to link your job to input parameters. You are required to create an association from the job entity or have Job as your input parameter and then cast it to the specified job in your microflow, which feels foolish because i already know the input type of the specific microflow.

* Fix to be able to find the Job Parameter

* Moved Core.getMetaObject to MicroflowRepository

* Fixed unit tests

* Changed readme

Co-authored-by: Menno de Haas <menno.dehaas@webflight.nl>